Update dependencies for Java21 and rename to demo/ to server/
