+++ src/main/java/io/anserini/server/Application.java
@@ -1,4 +1,4 @@
-package io.anserini.demo;
+package io.anserini.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
