+++ src/main/java/io/anserini/server/Controller.java
@@ -1,4 +1,4 @@
-package io.anserini.demo;
+package io.anserini.server;
 
 import java.util.List;
