+++ src/main/java/io/anserini/server/QueryResult.java
@@ -1,4 +1,4 @@
-package io.anserini.demo;
+package io.anserini.server;
 
 public class QueryResult {
   private String docid;
